+++ packages/@webex/plugin-meetings/test/unit/spec/meeting/index.js
@@ -3786,6 +3786,7 @@ describe('plugin-meetings', () => {
         it('should have #requestScreenShareFloor', () => {
           assert.exists(meeting.requestScreenShareFloor);
         });
+
         beforeEach(() => {
           meeting.locusInfo.mediaShares = [{name: 'content', url: url1}];
           meeting.locusInfo.self = {url: url1};
@@ -3793,13 +3794,47 @@ describe('plugin-meetings', () => {
           meeting.mediaProperties.shareVideoStream = {};
           meeting.mediaProperties.mediaDirection.sendShare = true;
           meeting.state = 'JOINED';
+          meeting.localShareInstanceId = '1234-5678';
         });
+
+        afterEach(() => {
+          sinon.restore();
+        });
+
         it('should send the share', async () => {
           const share = meeting.requestScreenShareFloor();
 
           assert.exists(share.then);
           await share;
           assert.calledOnce(meeting.meetingRequest.changeMeetingFloor);
+
+          assert.calledWith(webex.internal.newMetrics.submitClientEvent, {
+            name: 'client.share.floor-grant.request',
+            payload: {mediaType: 'share', shareInstanceId: '1234-5678'},
+            options: {meetingId: meeting.id},
+          });
+        });
+
+        it('should submit expected metric on failure', async () => {
+          const error = new Error('forced');
+
+          meeting.meetingRequest.changeMeetingFloor = sinon.stub().returns(Promise.reject(error));
+          const getChangeMeetingFloorErrorPayloadSpy = sinon
+            .stub(MeetingUtil, 'getChangeMeetingFloorErrorPayload')
+            .returns('foo');
+
+          await meeting.requestScreenShareFloor().catch((err) => {
+            assert.equal(err, error);
+          });
+
+          assert.calledWith(getChangeMeetingFloorErrorPayloadSpy, 'forced');
+
+          // ensure the expected CA share metric is submitted
+          assert.calledWith(webex.internal.newMetrics.submitClientEvent, {
+            name: 'client.share.floor-granted.local',
+            payload: {mediaType: 'share', errors: 'foo', shareInstanceId: '1234-5678'},
+            options: {meetingId: meeting.id},
+          });
         });
       });
 
@@ -5777,16 +5812,22 @@ describe('plugin-meetings', () => {
           const checkScreenShareVideoPublished = (stream) => {
             assert.calledOnce(meeting.requestScreenShareFloor);
 
-            assert.calledWith(
-              meeting.sendSlotManager.getSlot(MediaType.VideoSlides).publishStream,
-              stream
-            );
-            assert.equal(meeting.mediaProperties.shareVideoStream, stream);
+            // ensure the CA share metrics are submitted
+            assert.calledWith(webex.internal.newMetrics.submitClientEvent, {
+              name: 'client.share.initiated',
+              payload: {mediaType: 'share', shareInstanceId: meeting.localShareInstanceId},
+              options: {meetingId: meeting.id},
+            });
             assert.equal(meeting.mediaProperties.mediaDirection.sendShare, true);
           };
 
           const checkScreenShareAudioPublished = (stream) => {
-            assert.calledOnce(meeting.requestScreenShareFloor);
+            // ensure the CA share metrics are submitted
+            assert.calledWith(webex.internal.newMetrics.submitClientEvent, {
+              name: 'client.share.initiated',
+              payload: {mediaType: 'share', shareInstanceId: meeting.localShareInstanceId},
+              options: {meetingId: meeting.id},
+            });
 
             assert.calledWith(
               meeting.sendSlotManager.getSlot(MediaType.AudioSlides).publishStream,
@@ -9028,6 +9069,13 @@ describe('plugin-meetings', () => {
             assert.exists(whiteboardShare.then);
             await whiteboardShare;
             assert.calledOnce(meeting.meetingRequest.changeMeetingFloor);
+
+            // ensure the CA share metric is submitted
+            assert.calledWith(webex.internal.newMetrics.submitClientEvent, {
+              name: 'client.share.initiated',
+              payload: {mediaType: 'whiteboard'},
+              options: {meetingId: meeting.id},
+            });
           });
         });
         describe('#stopWhiteboardShare', () => {