+++ CHANGELOG.md
@@ -2,6 +2,41 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.0.0-beta.364](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.363...v3.0.0-beta.364) (2024-03-01)
+
+
+### Bug Fixes
+
+* **meetings:** deploy-pr ([#3422](https://github.com/webex/webex-js-sdk/issues/3422)) ([5ff424e](https://github.com/webex/webex-js-sdk/commit/5ff424ee723ea56b5f5101d6d34b4d0aebdbd9e0))
+
+## [3.0.0-beta.363](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.362...v3.0.0-beta.363) (2024-02-29)
+
+
+### Bug Fixes
+
+* **plugin-meetings-samples:** fix remote screen re-share ([#3410](https://github.com/webex/webex-js-sdk/issues/3410)) ([ab4a9a2](https://github.com/webex/webex-js-sdk/commit/ab4a9a2f861ff16fec93938870c71e7401fd4792))
+
+## [3.0.0-beta.362](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.361...v3.0.0-beta.362) (2024-02-28)
+
+
+### Features
+
+* **ca-metrics:** CA Share Error Metric ([#3170](https://github.com/webex/webex-js-sdk/issues/3170)) ([2b1be7e](https://github.com/webex/webex-js-sdk/commit/2b1be7ebfd591486ed7a55b21a7b7caa15f2d7a4))
+
+## [3.0.0-beta.361](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.360...v3.0.0-beta.361) (2024-02-28)
+
+
+### Bug Fixes
+
+* **test-users:** test user set preferred site docs ([#3411](https://github.com/webex/webex-js-sdk/issues/3411)) ([b5cb7ad](https://github.com/webex/webex-js-sdk/commit/b5cb7ad023069ae5637c6f97be0eebf8dc2e8063))
+
+## [3.0.0-beta.360](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.359...v3.0.0-beta.360) (2024-02-27)
+
+
+### Bug Fixes
+
+* don't do TCP reachability on VMN, add API to toggle TCP reachability ([#3401](https://github.com/webex/webex-js-sdk/issues/3401)) ([9a6a31e](https://github.com/webex/webex-js-sdk/commit/9a6a31e8e43ecd96b42f5d50564bc3c5664d6b53))
+
 ## [3.0.0-beta.359](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.358...v3.0.0-beta.359) (2024-02-26)